@@ -67,10 +67,9 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.34.0",
-    "@types/eslint__js": "^8.42.3",
     "@types/jest": "^29.5.12",
     "@types/node": "^24.3.0",
-    "eslint": "^8.57.0",
+    "eslint": "^9.34.0",
     "eslint-config-prettier": "^9.1.0",
     "husky": "^9.0.11",
     "jest": "^30.0.5",
@@ -79,7 +78,7 @@
     "prettier": "^3.3.2",
     "ts-jest": "^29.2.0",
     "typescript": "^5.9.2",
-    "typescript-eslint": "^7.16.0",
+    "typescript-eslint": "^8.41.0",
     "wireit": "^0.14.12"
   },
   "husky": {
